backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
drf-yasg==1.21.10
Pillow==10.4.0
PyYAML==6.0.2
